Refactor player handling and graphics notification functions
- Moved several functions related to player movement and graphics notifications from handle_player.c to new files: handle_player_ext2.c, handle_player_ext3.c, handle_player_ext4.c, and handle_player_ext5.c for better organization.
- Introduced helper functions for coordinate updates based on player direction.
- Consolidated resource management functions for tile elevation and player updates.
- Updated server.h to declare new functions for handling player actions and graphics notifications.
- Removed redundant code and improved readability by separating concerns into distinct files.
